include side bar on character sheet
